Fix(Multer): Fix upload finalizer return string
Description:
In the old finalizer function there was a problem in returned string
which would cause the file's final path to start from public directory
instead of upload directory. Since the public directory is supposed to
be static, then the final path shouldn't be started from public
directory.

Logs:
[Fixed][Common/Utils]: Fix multer upload finalizer function returned
value.