Feat(Decorators): Add an authorization skipping decorator
Description:
Some of the endpoints in the application should be accessible without
the authorization requirement even thought the other endpoints in the
same controller does. In order to achieve this a new decorator has been
added to the system. This decorator will set a true value in the
request's metadata which will tell the authorization guard to skip the
request.

Logs:
[Added][Common/Decorators]: Add auth skipping decorator.
[Modified][AuthModule]: Update theb auth guard to handle the skip
process.